Handle expected and unexpected http error statuses
Fixes #11
Auth errors are reported via HTTP status rather than json-rpc codes,
so we need to check the resty response for those.  Additionally we
assume that there could be other unexpected cases where an http error
status is returned without json-rpc schema and so we return a generic
service error for those.